backend/recommendation-system/requirements.txt
numpy>=1.24.3
scikit-learn>=1.3.0
fastapi>=0.68.1
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0 